implemented optimization of min (and max) floor
the optimization excludes cases in which all items are above at or above floor `i` and one item is taken to floor `i - 1` which is clearly a waste of moves. same applies for configurations coming from the end state.
the optimization seems to be almost ineffective, with a total of very few cases pruned and the same runtime cost as the version without this optimization.
i'm quite sure that the check for invalid moves can be made using only the state of the elevator